+++ app/src/main/java/com/anshyeon/onoff/data/repository/PostRepository.kt
@@ -1,5 +1,7 @@
 package com.anshyeon.onoff.data.repository
 
+import com.anshyeon.onoff.data.dataSource.ImageDataSource
+import com.anshyeon.onoff.data.dataSource.UserDataSource
 import com.anshyeon.onoff.data.model.ImageContent
 import com.anshyeon.onoff.data.model.Post
 import com.anshyeon.onoff.data.model.User
@@ -10,19 +12,17 @@ import com.anshyeon.onoff.network.extentions.onSuccess
 import com.anshyeon.onoff.network.model.ApiResponse
 import com.anshyeon.onoff.network.model.ApiResultException
 import com.anshyeon.onoff.util.DateFormatText
-import com.google.firebase.storage.FirebaseStorage
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.onCompletion
-import kotlinx.coroutines.tasks.await
 import javax.inject.Inject
 
 class PostRepository @Inject constructor(
     private val fireBaseApiClient: FireBaseApiClient,
     private val userDataSource: UserDataSource,
+    private val imageDataSource: ImageDataSource,
 ) {
 
     suspend fun createPost(
@@ -34,7 +34,7 @@ class PostRepository @Inject constructor(
     ): ApiResponse<Map<String, String>> {
         val currentTime = System.currentTimeMillis()
         val postId = userDataSource.getUid() + currentTime
-        val imageLocations = uploadImages(imageList)
+        val imageLocations = imageDataSource.uploadImages(imageList)
         val post = Post(
             postId,
             title,
@@ -69,7 +69,7 @@ class PostRepository @Inject constructor(
                     entry.value.run {
                         copy(
                             imageUrlList = imageLocations?.map { location ->
-                                getDownloadUrl(location)
+                                imageDataSource.downloadImage(location)
                             } ?: emptyList()
                         )
                     }
@@ -85,24 +85,4 @@ class PostRepository @Inject constructor(
     }.onCompletion {
         onComplete()
     }.flowOn(Dispatchers.Default)
-
-    private suspend fun getDownloadUrl(location: String): String {
-        return FirebaseStorage.getInstance()
-            .getReference(location)
-            .downloadUrl
-            .await()
-            .toString()
-    }
-
-    private suspend fun uploadImages(imageList: List<ImageContent>): List<String> = coroutineScope {
-        imageList.map { image ->
-            val storageRef = FirebaseStorage.getInstance().reference
-            val location = "images/${image.fileName}"
-            val imageRef = storageRef.child(location)
-            imageRef
-                .putFile(image.uri)
-                .await()
-            location
-        }
-    }
 }
